pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer - analyzes the specified URLs for SEO suitability."
authors = ["Sergey Chapurin <goodwin7@list.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
classifiers = [
  "Topic :: Education",
  "Programming Language :: Python :: 3.10",
  "Framework :: Flask",
  "Environment :: Web Environment",
  "License :: Freeware",
]


[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.33.0"
requests = "^2.32.3"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
